infoDoRNG {doRNG}	R Documentation

Getting Information About doRNG Foreach Backend

Description

infoDoRNG returns information about the doRNG backend, e.g., version, number of workers. It is not meant to be called by the user.

Usage

infoDoRNG(data, item)

doRNG(obj, ex, envir, data)

Arguments

data
configuration data of the doRNG backend

item
the data item requested, as a character string (e.g. 'name', 'workers', 'version')

obj
a foreach description of the loop arguments

ex
the lopp expression

envir
the loop's evaluation environment

Value

infoDoRNG returns the requested info (usually as a character string or a numeric value).

Functions

doRNG(): implements the generic reproducible foreach backend. It should not be called directly by the user.

Author(s)

Renaud Gaujoux

[Package doRNG version 1.8.6.3 Index]